process_document_analysis/requirements.txt
boto3
requests
aws-psycopg2==1.3.8
